test: misc test improvements (#9812)
* more opts

* ignore can_disable_block_gas_limit

* move selector tests to integration tests

* reduce sleep

* rm unnecessary test, already tested with openchain/4byte commands

* rm needless_return allows

* chore: clippy

* move gas_report_fuzz_invariant to integration
