+++ crates/anvil/src/eth/backend/mem/cache.rs
@@ -47,7 +47,7 @@ impl DiskStateCache {
                 }
             }
         }
-        if let Some(ref temp_dir) = self.temp_dir {
+        if let Some(temp_dir) = &self.temp_dir {
             let path = temp_dir.path().join(format!("{hash:?}.json"));
             Some(f(path))
         } else {
