+++ crates/cast/bin/main.rs
@@ -57,7 +57,6 @@ fn run() -> Result<()> {
     main_args(args)
 }
 
-#[allow(clippy::needless_return)]
 #[tokio::main]
 async fn main_args(args: CastArgs) -> Result<()> {
     match args.cmd {
