+++ crates/common/src/selectors.rs
@@ -2,11 +2,10 @@
 
 #![allow(missing_docs)]
 
-use crate::abi::abi_decode_calldata;
+use crate::{abi::abi_decode_calldata, provider::runtime_transport::RuntimeTransportBuilder};
 use alloy_json_abi::JsonAbi;
 use alloy_primitives::map::HashMap;
 use eyre::Context;
-use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use std::{
     fmt,
@@ -16,6 +15,8 @@ use std::{
     },
     time::Duration,
 };
+
+const BASE_URL: &str = "https://api.openchain.xyz";
 const SELECTOR_LOOKUP_URL: &str = "https://api.openchain.xyz/signature-database/v1/lookup";
 const SELECTOR_IMPORT_URL: &str = "https://api.openchain.xyz/signature-database/v1/import";
 
@@ -38,20 +39,17 @@ pub struct OpenChainClient {
 }
 
 impl OpenChainClient {
-    /// Creates a new client with default settings
+    /// Creates a new client with default settings.
     pub fn new() -> eyre::Result<Self> {
-        let inner = reqwest::Client::builder()
-            .default_headers(HeaderMap::from_iter([(
-                HeaderName::from_static("user-agent"),
-                HeaderValue::from_static("forge"),
-            )]))
-            .timeout(REQ_TIMEOUT)
+        let inner = RuntimeTransportBuilder::new(BASE_URL.parse().unwrap())
+            .with_timeout(REQ_TIMEOUT)
             .build()
+            .reqwest_client()
             .wrap_err("failed to build OpenChain client")?;
         Ok(Self {
             inner,
-            spurious_connection: Arc::new(Default::default()),
-            timedout_requests: Arc::new(Default::default()),
+            spurious_connection: Default::default(),
+            timedout_requests: Default::default(),
             max_timedout_requests: MAX_TIMEDOUT_REQ,
         })
     }
@@ -577,62 +575,11 @@ pub fn parse_signatures(tokens: Vec<String>) -> ParsedSignatures {
 }
 
 #[cfg(test)]
-#[allow(clippy::disallowed_macros)]
-#[allow(clippy::needless_return)]
 mod tests {
     use super::*;
 
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_decode_selector() {
-        let sigs = decode_function_selector("0xa9059cbb").await;
-        assert_eq!(sigs.unwrap()[0], "transfer(address,uint256)".to_string());
-
-        let sigs = decode_function_selector("a9059cbb").await;
-        assert_eq!(sigs.unwrap()[0], "transfer(address,uint256)".to_string());
-
-        // invalid signature
-        decode_function_selector("0xa9059c")
-            .await
-            .map_err(|e| {
-                assert_eq!(
-                    e.to_string(),
-                    "Invalid selector 0xa9059c: expected 10 characters (including 0x prefix)."
-                )
-            })
-            .map(|_| panic!("Expected fourbyte error"))
-            .ok();
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_decode_calldata() {
-        let decoded = decode_calldata("0xa9059cbb0000000000000000000000000a2ac0c368dc8ec680a0c98c907656bd970675950000000000000000000000000000000000000000000000000000000767954a79").await;
-        assert_eq!(decoded.unwrap()[0], "transfer(address,uint256)".to_string());
-
-        let decoded = decode_calldata("a9059cbb0000000000000000000000000a2ac0c368dc8ec680a0c98c907656bd970675950000000000000000000000000000000000000000000000000000000767954a79").await;
-        assert_eq!(decoded.unwrap()[0], "transfer(address,uint256)".to_string());
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_import_selectors() {
-        let mut data = RawSelectorImportData::default();
-        data.function.push("transfer(address,uint256)".to_string());
-        let result = import_selectors(SelectorImportData::Raw(data)).await;
-        assert_eq!(
-            result.unwrap().result.function.duplicated.get("transfer(address,uint256)").unwrap(),
-            "0xa9059cbb"
-        );
-
-        let abi: JsonAbi = serde_json::from_str(r#"[{"constant":false,"inputs":[{"name":"_to","type":"address"},{"name":"_value","type":"uint256"}],"name":"transfer","outputs":[{"name":"","type":"bool"}],"payable":false,"stateMutability":"nonpayable","type":"function", "methodIdentifiers": {"transfer(address,uint256)(uint256)": "0xa9059cbb"}}]"#).unwrap();
-        let result = import_selectors(SelectorImportData::Abi(vec![abi])).await;
-        println!("{result:?}");
-        assert_eq!(
-            result.unwrap().result.function.duplicated.get("transfer(address,uint256)").unwrap(),
-            "0xa9059cbb"
-        );
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_parse_signatures() {
+    #[test]
+    fn test_parse_signatures() {
         let result = parse_signatures(vec!["transfer(address,uint256)".to_string()]);
         assert_eq!(
             result,
@@ -685,60 +632,6 @@ mod tests {
         assert_eq!(
             result,
             ParsedSignatures { signatures: Default::default(), ..Default::default() }
-        );
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_decode_event_topic() {
-        let decoded = decode_event_topic(
-            "0x7e1db2a1cd12f0506ecd806dba508035b290666b84b096a87af2fd2a1516ede6",
-        )
-        .await;
-        assert_eq!(decoded.unwrap()[0], "updateAuthority(address,uint8)".to_string());
-
-        let decoded =
-            decode_event_topic("7e1db2a1cd12f0506ecd806dba508035b290666b84b096a87af2fd2a1516ede6")
-                .await;
-        assert_eq!(decoded.unwrap()[0], "updateAuthority(address,uint8)".to_string());
-
-        let decoded = decode_event_topic(
-            "0xb7009613e63fb13fd59a2fa4c206a992c1f090a44e5d530be255aa17fed0b3dd",
-        )
-        .await;
-        assert_eq!(decoded.unwrap()[0], "canCall(address,address,bytes4)".to_string());
-
-        let decoded =
-            decode_event_topic("b7009613e63fb13fd59a2fa4c206a992c1f090a44e5d530be255aa17fed0b3dd")
-                .await;
-        assert_eq!(decoded.unwrap()[0], "canCall(address,address,bytes4)".to_string());
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_decode_selectors() {
-        let event_topics = vec![
-            "7e1db2a1cd12f0506ecd806dba508035b290666b84b096a87af2fd2a1516ede6",
-            "0xb7009613e63fb13fd59a2fa4c206a992c1f090a44e5d530be255aa17fed0b3dd",
-        ];
-        let decoded = decode_selectors(SelectorType::Event, event_topics).await;
-        let decoded = decoded.unwrap();
-        assert_eq!(
-            decoded,
-            vec![
-                Some(vec!["updateAuthority(address,uint8)".to_string()]),
-                Some(vec!["canCall(address,address,bytes4)".to_string()]),
-            ]
-        );
-
-        let function_selectors = vec!["0xa9059cbb", "0x70a08231", "313ce567"];
-        let decoded = decode_selectors(SelectorType::Function, function_selectors).await;
-        let decoded = decoded.unwrap();
-        assert_eq!(
-            decoded,
-            vec![
-                Some(vec!["transfer(address,uint256)".to_string()]),
-                Some(vec!["balanceOf(address)".to_string()]),
-                Some(vec!["decimals()".to_string()]),
-            ]
         );
     }
 }
