+++ crates/evm/traces/src/identifier/signatures.rs
@@ -179,40 +179,3 @@ impl Drop for SignaturesIdentifier {
         self.save();
     }
 }
-
-#[cfg(test)]
-#[allow(clippy::needless_return)]
-mod tests {
-    use super::*;
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn can_query_signatures() {
-        let tmp = tempfile::tempdir().unwrap();
-        {
-            let sigs = SignaturesIdentifier::new(Some(tmp.path().into()), false).unwrap();
-
-            assert!(sigs.read().await.cached.events.is_empty());
-            assert!(sigs.read().await.cached.functions.is_empty());
-
-            let func = sigs.write().await.identify_function(&[35, 184, 114, 221]).await.unwrap();
-            let event = sigs
-                .write()
-                .await
-                .identify_event(&[
-                    39, 119, 42, 220, 99, 219, 7, 170, 231, 101, 183, 30, 178, 181, 51, 6, 79, 167,
-                    129, 189, 87, 69, 126, 27, 19, 133, 146, 216, 25, 141, 9, 89,
-                ])
-                .await
-                .unwrap();
-
-            assert_eq!(func, get_func("transferFrom(address,address,uint256)").unwrap());
-            assert_eq!(event, get_event("Transfer(address,address,uint128)").unwrap());
-
-            // dropping saves the cache
-        }
-
-        let sigs = SignaturesIdentifier::new(Some(tmp.path().into()), false).unwrap();
-        assert_eq!(sigs.read().await.cached.events.len(), 1);
-        assert_eq!(sigs.read().await.cached.functions.len(), 1);
-    }
-}
