+++ crates/forge/bin/cmd/doc/server.rs
@@ -90,7 +90,6 @@ impl Server {
     }
 }
 
-#[allow(clippy::needless_return)]
 #[tokio::main]
 async fn serve(build_dir: PathBuf, address: SocketAddr, file_404: &str) -> io::Result<()> {
     let file_404 = build_dir.join(file_404);
