+++ crates/forge/bin/cmd/soldeer.rs
@@ -40,7 +40,6 @@ mod tests {
     use soldeer_commands::{commands::Version, Command};
 
     #[tokio::test]
-    #[allow(clippy::needless_return)]
     async fn test_soldeer_version() {
         let command = Command::Version(Version::default());
         assert!(soldeer_commands::run(command).await.is_ok());
