+++ crates/forge/tests/it/main.rs
@@ -1,5 +1,3 @@
-#![allow(clippy::needless_return)]
-
 pub mod config;
 pub mod test_helpers;
